+++ tests/test_spyna.py
@@ -122,5 +122,21 @@ def test_attr_filter():
     check("foo.bar.baz[a=2]")
 
 
+def test_filter_group():
+    check("foo[bar, baz]")
+
+
 def test_null():
     check("null")
+
+
+def test_array():
+    check("[1, 'a']")
+
+
+def test_select_all():
+    check("select(*)")
+
+
+def test_group():
+    check("foo, bar, baz")